Move data to JSON file
Move the collection of recipes from DynamoDB to a JSON file, so that
it is easier to edit and maintain. I'm also kind of tired of working
on web services and databases, so this will let me focus on the front
end, which I'm more excited about doing at the moment.