Remove background sync
Remove the background sync and instead sync the recipes with the server
when the application first loads. This way, it will guarantee to have
the latest list of recipes each time the app starts. The current
implementation has a bug that causes it to fail to sync correctly. The
value of trying to use the background sync is too low compared to the
extra complexity it adds.