Flatten the project hierarchy
Move /data/recipes to /recipes. The recipes subdirectory was the only
directory under /data. There is no reason to have the extra
subdirectory.

Move the React web app from /client to the root of the project.

Move the build scripts (there was only one) from /scripts to the root of
the project.